backend/requirements.txt
fastapi
uvicorn
youtube-transcript-api
httpx
google-genai
dotenv